data/docker/requirements.txt
numpy
matplotlib
OpenVisus
pyyaml
scipy